requirements.txt
astroid==2.2.5
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
pdf2image==1.5.4
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.1
